requirements.txt
torch
torchvision
timm
numpy
pandas
matplotlib
scikit-learn
onnx
onnxruntime
gradio
jupyter
Pillow
